backend/src/requirements.txt
Django==2.1.2
pyodbc
django-pyodbc-azure
djangorestframework==3.8.2
markdown==3.0.1
django-filter==2.0.0